feat: ProcessViewController의 setDiaryWeatherInformation 메서드 추가
